can retrieve list of recipes from api
