+++ app/controllers/recipes_controller.rb
@@ -2,5 +2,7 @@ class RecipesController < ApplicationController
 
   def index
     @query = params[:q]
+    @recipes = ApiMuncherWrapper.get_recipes(@query)
   end
+
 end
